requirements.txt
werkzeug==2.0.1
opencv-python==4.5.3.56
ultralytics==8.0.20
torch==2.0.1
torchvision==0.15.2
numpy==1.21.0
pillow==8.3.1
tqdm==4.62.0
pyyaml==6.0
pymongo==4.3.3
dnspython==2.3.0
pytesseract==0.3.10
